Add dependency of FIP to router interface.
On CityCloud, I get
| stack_status          | CREATE_FAILED
| stack_status_reason   | Resource CREATE failed: NotFound:
  resources.manager_floating_ip: External network 2aec7a99-3783-4e2a-bd2b-bbe4fef97d1c is not reachable from subnet fff93f95-d82c-4e7b-81cd-f2c76e37c1ce.  Therefore, cannot associate Port 8d753df3-d569-481f-8438-ac 34f9f94b27 with a Floating IP.

The manager floating ip needs to be associated with a fixed IP address
that is L3 reachable from the public network that we allocate the FIP
from. neutron rightly enforces it, the heat template correctly attaches
the subnet to the router but it misses the dependency, so we hit a race.

Adding a depency fixes it and makes the deployment succeed on CityCloud.

Signed-off-by: Kurt Garloff <scs@garloff.de>