enable_snat: true for router
Some OpenStack deployments chose to not default to enable snat for their
routers. OTC is one example. While unusual, it's a valid choice.
This leads to instance not being able to access the internet which makes
the deployment fail.

enable_snat: true for the router takes care of it. It's harmless on
clouds where it's enabled by default.

Signed-off-by: Kurt Garloff <scs@garloff.de>