Specify availability_zone for volumes.
In some clouds, there is more than one AZ for volumes, so it makes
sense to specify it. It typically should match the AZ for the VMs,
but let's keep it separate for flexibility. (Some clouds may have
several AZs for VMs, but only one replicated solution across AZs.)
In OTC, you need to match volume_az with availabilty_zone, otherwise
attachment fails. Leaving it open leaves it to chance ...

Signed-off-by: Kurt Garloff <scs@garloff.de>