Add parameter drives_vdx to use vdX drives names.
This will be required on many clouds. Some clouds to virtio-scsi with
sdX names, some do virtio-blk with vdX names (and some folks might even
have other virtual drive names such as xvdX on Xen ...).
We will need a mroe generic mechanism here.
Rather than making this very flexible, we have a reasonable chance to
autodetect this, so the next iteration will attempt to do so.

Signed-off-by: Kurt Garloff <scs@garloff.de>