Remove virtualbmc service (#1614)
The virtualbmc service was included as a workaround to test Ironic
for functionality within the testbed. In the future we will have access
to real baremetal systems to perform real tests of Ironic. Therefore,
the approach to use emulated baremetal systems within the testbed will
not be pursued further. The deployment of Ironic continues to be tested
in the testbed.

Signed-off-by: Christian Berendt <berendt@osism.tech>