Configure chrony with cloud-init
Sometimes it happens that the system time is not correct after a boot.
This then leads to anomalies in the bootstrap process. For example, the
mtime on certain files is incorrect as a result. To work around the
problem in the testbed, Chrony is already deployed with cloud-init.
Thus a correct system time is available directly after the boot even
before the bootstrap process was executed.

$ stat /var/lib/apt/periodic/update-success-stamp
  File: /var/lib/apt/periodic/update-success-stamp
  Size: 0         	Blocks: 0          IO Block: 4096   regular empty file
Device: 801h/2049d	Inode: 69167       Links: 1
Access: (0644/-rw-r--r--)  Uid: (    0/    root)   Gid: (    0/    root)
Access: 2023-05-12 10:26:37.609383906 +0000
Modify: 2023-05-12 10:36:46.308000000 +0000
Change: 2023-05-12 10:36:46.308000000 +0000
 Birth: 2023-04-28 02:10:21.728204937 +0000

Signed-off-by: Christian Berendt <berendt@osism.tech>