Pin docker_cli_version
The Docker CLI version must also be pinned as the default of the Docker CLI
version was only recently set to the Docker version itself. If the Docker CLI
version has not been pinned, problems may occur if only the Docker CLI is
downgraded/upgraded after the initial deployment of Docker.

Signed-off-by: Christian Berendt <berendt@osism.tech>