terraform: avoid deleting attached FIPs, better port filtering (#2419)
The filter in the search_floating_ip() statement does not do what it is
supposed to. Instead skip if there is a fixed_ip_address.

For port cleanup, provide better logic:
* Delete port unconditionally if it's name matches the prefix
* Keep it unconditionally if it has a name that does not match
* IF we have a list of IPADDR strings the beginning of the ip
  addresses will be matched against the IPADDR strings; delete
  if there is a match, do not delete if there is no match.
  Always match, if no IPADDR filter is passed.

The filtering can be used to ensure we only delete IP addresses from
IP ranges in subnets that we have created. This improves coexistence
with other infrastructure in the smae project.
(If you can, avoid coexistence, though.)

This is a port from
https://github.com/SovereignCloudStack/standards/pull/712

Also simplify exception handling when looking for OS_CLOUD.

Signed-off-by: Kurt Garloff <kurt@garloff.de>
Co-authored-by: Matthias Büchse <github@mbue.de>