Add baremetal deployment
Add a deployment scenario, where `testbed-resource-nodes` are used to
provide virtual machines for baremetal deployment.
Virtual baremetal machines are created using `tenks` ([1]), which may
later also be used to register baremetal nodes and create matching
flavors.
The network `vxlan0`, which serves as an external network in ordinary
deployments is used as a provider network conected to the virtual
baremetal. Additionally an out-of-band network is created to manage the
machines via virtual BMC.
Since routing to the API network did not seem to work during PXE
booting, ports are created for the ironic tftp and httpd services and
connected to the control plane in a similar way to what is done in kolla
for octavia amphora management networks ([2]). Routing to the ironic API
during callback of the ironic python agent image works fine and is done
via the manager node.

[1]
https://docs.openstack.org/tenks

[2]
https://opendev.org/openstack/kolla-ansible/src/branch/master/ansible/roles/octavia/tasks/hm-interface.yml

Part of https://github.com/osism/issues/issues/1196
Depends on https://github.com/osism/testbed/pull/2606

Signed-off-by: Jan Horstmann <horstmann@osism.tech>