Use network role to create vxlans
Use the network role to create vxlans. This is mostly a drop-in
replacement with the exception that it does not add IP addresses to the
`vxlan0` interface except for the manager node. Since the vxlan is used
as a neutron external network attached to an openvswitch bridge on all
nodes except for the manager, the address is unusable there anyway.

Signed-off-by: Jan Horstmann <horstmann@osism.tech>