Do not restart old manager service
Do not restart the old manager service `docker-compose@manager.service`
on osism versions using the new `manager.service`.

This fixes a bug where applying the manager role would stop the manager
service when trying to stop and disable the old manager service, which
would be active due to the restart and issue a `docker-compose down`.

Signed-off-by: Jan Horstmann <horstmann@osism.tech>